Switch the README.md instructions to 0.2.0 since we're releasing it imminently.
If I am going to do this each release, I will script (maybe a pre-commit
hook that looks at the meson.build project version and then substitutes
that into the README).  However, I might not update it always since the
instructions say to run pre-commit autoupdate to update the rev.  It's
just that the 0.1.0 release is unusable due to the meson, ninja, and
mesonpep517 interdependency problems.