Fix the badge links
I created this with captain_hooks as the slug in cookiecutter because it
wanted something that was a valid python identifier but pushed it as
captain-hooks when I made the git repo because I like dashes better in
project names.  I should look into separating those out in the
cookiecutter template later.